added authorization to gigRouter endpoints

+++ routes/gigRouter.js
@@ -1,6 +1,6 @@
-
 const express = require('express');
 const Gig =  require('../models/gigs');
+const authenticate = require('../authenticate');
 
 const gigRouter = express.Router();
 
@@ -14,7 +14,9 @@ gigRouter.route('/')
     })
     .catch(err => next(err));
 })
-.post((req, res, next) => {
+.post(authenticate.verifyUser, (req, res, next) => {
+    console.log(req.user);
+    req.body.author = req.user._id;
     Gig.create(req.body)
     .then(gig => {
         console.log('Gig Created ', gig);
@@ -24,11 +26,11 @@ gigRouter.route('/')
     })
     .catch(err => next(err));
 })
-.put((req, res) => {
+.put(authenticate.verifyUser, (req, res) => {
     res.statusCode = 405;
     res.end('PUT operation not supported on /gigs');
 })
-.delete((req, res, next) => {
+.delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
     Gig.deleteMany()
     .then(response => {
         res.statusCode = 200;
@@ -48,27 +50,68 @@ gigRouter.route('/:gigId')
     })
     .catch(err => next(err));
 })
-.post((req, res) => {
+.post(authenticate.verifyUser, (req, res) => {
     res.statusCode = 405;
     res.end(`POST operation not supported on /gigs/${req.params.gigId}`);
 })
-.put((req, res, next) => {
-    Gig.findByIdAndUpdate(req.params.gigId, {
-        $set: req.body
-    }, { new: true })
+//Need to update this PUT so only the user who created the gig can update it.
+.put(authenticate.verifyUser, (req, res, next) => {
+    console.log(req.user);
+    Gig.findById(req.params.gigId)
+    .populate('author')
     .then(gig => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(gig);
+        if (gig) {
+            console.log(gig);
+            console.log(typeof gig);
+            if (req.user._id.equals(gig.author._id)) {
+                Gig.findByIdAndUpdate(req.params.gigId, {  
+                    $set: req.body
+                }, { new: true })
+                .then(gig => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(gig); 
+                })
+                .catch(err => next(err));
+            } else {
+                const err = new Error('Operation not authorized!');
+                err.status = 403;
+                return next(err);
+            }
+        } else {
+            const err = new Error(`Gig ${req.params.gigId} not found`);
+            err.status = 404;
+            return next(err);
+        }
     })
     .catch(err => next(err));
 })
-.delete((req, res, next) => {
-    Gig.findByIdAndDelete(req.params.gigId)
-    .then(response => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(response);
+.delete(authenticate.verifyUser, (req, res, next) => {
+    console.log(req.user);
+    Gig.findById(req.params.gigId)
+    .populate('author')
+    .then(gig => {
+        if (gig) {
+            console.log(gig);
+            console.log(typeof gig);
+            if (req.user._id.equals(gig.author._id)) {
+                Gig.findByIdAndDelete(req.params.gigId)
+                .then(response => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(response);
+                })
+                .catch(err => next(err));
+            } else {
+                const err = new Error('Operation not authorized!');
+                err.status = 403;
+                return next(err);
+            }
+        } else {
+            const err = new Error(`Gig ${req.params.gigId} not found`);
+            err.status = 404;
+            return next(err);
+        }
     })
     .catch(err => next(err));
 });
